requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
tensorflow==2.16.2
protobuf==4.25.3
numpy==1.26.3
matplotlib==3.8.2
networkx==3.3
pyyaml==6.0.1
scipy==1.14.0
scikit-learn==1.3.2
statsmodels==0.14.2
Pillow==10.4.0
netgraph==4.13.2
graphviz==0.20.3
pandas==2.1.4
six==1.16.0
requests==2.32.3
keras==3.4.1
pmdarima==2.0.4
tensorly==0.8.1
pytest==8.2.2
pygraphviz==1.13
gym==0.26.2
mujoco==3.1.6
gymnasium==0.29.1